dev-requirements.txt
# Formatters
black
isort

# Linters
dep-check
mypy
openapi-spec-validator
pre-commit
pyenchant
pylint
md-toc
yamllint

# Test packages
pytest
pytest-cov

# Tooling
ipdb
ipython
jinja2-cli
